Move nested types into their own files.
